requirements.txt
Django
requests
beautifulsoup4
celery
redis
djangorestframework
django-filter